Refactor Application::OnInit and organize source/ui
- Extract crash recovery, theme, and update logic from Application::OnInit to dedicated managers:
  - CrashRecoveryManager
  - ThemeManager
  - UpdateManager
- Move UI windows to `source/ui/windows/`
- Move UI dialogs to `source/ui/dialogs/`
- Update CMakeLists.txt and include paths
- Update includes in all source files
- Remove business logic from Application class (SRP)
