+++ source/app/managers/crash_recovery_manager.h
@@ -1,0 +1,9 @@
+#ifndef RME_APP_MANAGERS_CRASH_RECOVERY_MANAGER_H_
+#define RME_APP_MANAGERS_CRASH_RECOVERY_MANAGER_H_
+
+class CrashRecoveryManager {
+public:
+	static bool CheckAndRecover();
+};
+
+#endif
