+++ source/app/managers/theme_manager.h
@@ -1,0 +1,9 @@
+#ifndef RME_APP_MANAGERS_THEME_MANAGER_H_
+#define RME_APP_MANAGERS_THEME_MANAGER_H_
+
+class ThemeManager {
+public:
+	static void ApplyTheme();
+};
+
+#endif
