+++ source/palette/panels/brush_palette_panel.cpp
@@ -1,7 +1,7 @@
 #include "palette/panels/brush_palette_panel.h"
 #include "ui/gui.h"
-#include "ui/add_tileset_window.h"
-#include "ui/add_item_window.h"
+#include "ui/windows/add_tileset_window.h"
+#include "ui/windows/add_item_window.h"
 #include "game/materials.h"
 #include "palette/palette_window.h"
 #include "util/image_manager.h"
