+++ source/rendering/ui/map_display.cpp
@@ -35,7 +35,7 @@
 #include "game/item.h"
 #include "ui/properties/old_properties_window.h"
 #include "ui/properties/properties_window.h"
-#include "ui/tileset_window.h"
+#include "ui/windows/tileset_window.h"
 #include "palette/palette_window.h"
 #include "rendering/ui/screenshot_controller.h"
 #include "rendering/utilities/tile_describer.h"
@@ -50,7 +50,7 @@
 #include "app/application.h"
 #include "live/live_server.h"
 #include "live/live_client.h"
-#include "ui/browse_tile_window.h"
+#include "ui/windows/browse_tile_window.h"
 #include "ui/dialog_helper.h"
 #include "game/animation_timer.h"
 #include "ui/map_popup_menu.h"
