+++ source/ui/dialogs/map_statistics_dialog.h
