+++ source/ui/main_menubar.cpp
@@ -20,7 +20,7 @@
 #include "ui/main_menubar.h"
 #include "ui/dialog_util.h"
 #include "ui/gui.h"
-#include "ui/tool_options_window.h"
+#include "ui/windows/tool_options_window.h"
 #include "ui/tile_properties/tile_properties_panel.h"
 
 #include "map/map_statistics.h"
@@ -30,16 +30,16 @@
 #include "ui/map/towns_window.h"
 #include "ui/map/map_properties_window.h"
 
-#include "ui/about_window.h"
+#include "ui/windows/about_window.h"
 #include "ui/dat_debug_view.h"
 #include "ui/menubar_loader.h"
-#include "ui/map_statistics_dialog.h"
-#include "ui/live_dialogs.h"
+#include "ui/dialogs/map_statistics_dialog.h"
+#include "ui/dialogs/live_dialogs.h"
 
-#include "ui/extension_window.h"
-#include "ui/find_item_window.h"
+#include "ui/windows/extension_window.h"
+#include "ui/windows/find_item_window.h"
 #include "app/preferences.h"
-#include "ui/result_window.h"
+#include "ui/windows/result_window.h"
 
 #include "ui/menubar/search_handler.h"
 #include "ui/menubar/view_settings_handler.h"
