+++ source/ui/managers/welcome_manager.cpp
@@ -1,6 +1,6 @@
 #include "ui/managers/welcome_manager.h"
 #include "ui/gui.h"
-#include "ui/welcome_dialog.h"
+#include "ui/dialogs/welcome_dialog.h"
 #include "ui/main_menubar.h"
 #include "app/main.h"
