+++ source/ui/menubar/navigation_menu_handler.cpp
@@ -6,7 +6,7 @@
 #include "ui/map_tab.h"
 #include "ui/map_window.h"
 #include "ui/dialogs/goto_position_dialog.h"
-#include "ui/find_item_window.h"
+#include "ui/windows/find_item_window.h"
 #include "ui/main_menubar.h"
 #include "app/preferences.h"
