+++ source/ui/tile_properties/container_property_panel.cpp
@@ -9,7 +9,7 @@
 #include "map/tile.h"
 #include "map/map.h"
 #include "ui/dialog_util.h"
-#include "ui/find_item_window.h"
+#include "ui/windows/find_item_window.h"
 #include "ui/gui_ids.h"
 #include "util/image_manager.h"
 #include "ui/gui.h"
